requirements.txt
playwright==1.55.0
beautifulsoup4==4.12.2
mysql-connector-python==8.1.1
lxml==4.9.3
